Some updates to tests
