+++ docker/integration_tests/trim_integration/Dockerfile
@@ -5,4 +5,5 @@ WORKDIR tests
 COPY features .
 RUN pip install mysql-connector-python behave
 RUN ls
+
 ENTRYPOINT ["behave"]
